requirements.txt
python-telegram-bot==21.6
google-generativeai==0.8.3
google-genai==0.3.0      # Dedicated embeddings client
numpy==2.1.2
scikit-learn==1.5.2  # For cosine similarity calculation
httpx==0.27.2        # For modern, async HTTP requests
beautifulsoup4==4.12.3 # For parsing HTML from channel web previews
python-dotenv==1.1.1 # For environment variable management
aiofiles==23.2.1  # Async file I/O for persistence

pytest==7.4.4           # Testing framework (pinned for compatibility)
anyio==4.3.0            # Async compatibility layer
pytest-asyncio==0.21.1  # Async testing support
pytest-tornasync==0.6.0.post2  # Tornado async testing
pytest-trio==0.8.0      # Trio async testing
